gfs2_edit: Avoid a potential int overflow in dump_journal
Spotted by coverity: "Potentially overflowing expression
(rgd->bits->bi_start + rgd->bits->bi_len) * 4U with type unsigned int
(32 bits, unsigned) is evaluated using 32-bit arithmetic, and then used
in a context that expects an expression of type uint64_t (64 bits,
unsigned)"

+++ gfs2/edit/journal.c
@@ -530,7 +530,7 @@ void dump_journal(const char *journal, int tblk)
 				   rgd->ri.ri_length);
 			o = tblk - rgd->ri.ri_data0;
 			if (o >= (rgd->bits->bi_start +
-				  rgd->bits->bi_len) * GFS2_NBBY)
+				  rgd->bits->bi_len) * (uint64_t)GFS2_NBBY)
 				o += (sizeof(struct gfs2_rgrp) -
 				      sizeof(struct gfs2_meta_header))
 					* GFS2_NBBY;
